% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Hitting Binance API to get bitcoin price in USD}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{the no. of retries previously}
}
\description{
Hitting Binance API to get bitcoin price in USD
}
